DeveloperGuide/step-2-developer-journey-steps/06-04-FunctionApp-MSI/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
mysql-connector-python
azure.identity